refactor(dashboard): redesign sidebar to shadcn dashboard-01 pattern
- Create nav-main.tsx: data-driven navigation with NavGroup[] prop
- Create nav-secondary.tsx: secondary links (Docs, Home) with mt-auto
- Create nav-user.tsx: user dropdown in sidebar footer with Clerk integration
  (replaces both _sidebar-user display and inline topbar dropdown)
- Restructure app-sidebar.tsx to dashboard-01 composition pattern
- Move user menu from topbar to sidebar footer (NavUser component)
- Delete old _sidebar-nav.tsx, _sidebar-user.tsx, _sidebar-footer-links.tsx
