+++ packages/dashboard/src/components/sidebar/nav-secondary.tsx
@@ -1,0 +1,48 @@
+"use client";
+
+import type { LucideIcon } from "lucide-react";
+import Link from "next/link";
+import { usePathname } from "next/navigation";
+import {
+  SidebarGroup,
+  SidebarGroupContent,
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+} from "@/components/ui/sidebar";
+
+interface NavItem {
+  title: string;
+  url: string;
+  icon: LucideIcon;
+}
+
+export function NavSecondary({ items, className }: { items: NavItem[]; className?: string }) {
+  const pathname = usePathname();
+
+  return (
+    <SidebarGroup className={className}>
+      <SidebarGroupContent>
+        <SidebarMenu>
+          {items.map((item) => {
+            const isActive = item.url === "/" ? pathname === "/" : pathname.startsWith(item.url);
+            return (
+              <SidebarMenuItem key={item.url}>
+                <Link href={item.url}>
+                  <SidebarMenuButton
+                    className={`h-9 rounded-lg px-2.5 transition-colors ${isActive ? "bg-brand-soft text-brand font-medium" : "hover:bg-sidebar-accent"}`}
+                    isActive={isActive}
+                    tooltip={item.title}
+                  >
+                    <item.icon />
+                    <span>{item.title}</span>
+                  </SidebarMenuButton>
+                </Link>
+              </SidebarMenuItem>
+            );
+          })}
+        </SidebarMenu>
+      </SidebarGroupContent>
+    </SidebarGroup>
+  );
+}
