+++ packages/dashboard/src/components/sidebar/nav-user.tsx
@@ -1,0 +1,110 @@
+"use client";
+
+import { SignInButton, useAuth, useClerk, useUser } from "@clerk/react";
+import { ChevronsUpDownIcon, LogOutIcon, SettingsIcon, UserIcon } from "lucide-react";
+import Link from "next/link";
+import { ThemeToggle } from "@/components/theme-toggle";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import {
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  useSidebar,
+} from "@/components/ui/sidebar";
+
+export function NavUser() {
+  const { isSignedIn, isLoaded } = useAuth();
+  const { user } = useUser();
+  const { signOut } = useClerk();
+  const { isMobile } = useSidebar();
+
+  if (!isLoaded) return null;
+
+  if (!isSignedIn) {
+    return (
+      <SidebarMenu>
+        <SidebarMenuItem>
+          <SignInButton>
+            <button
+              type="button"
+              className="flex h-8 w-full items-center justify-center rounded-lg border border-border text-xs font-medium transition-colors hover:bg-accent"
+            >
+              Sign in
+            </button>
+          </SignInButton>
+        </SidebarMenuItem>
+      </SidebarMenu>
+    );
+  }
+
+  const initial =
+    user?.firstName?.[0] || user?.emailAddresses?.[0]?.emailAddress?.[0]?.toUpperCase() || "?";
+
+  return (
+    <DropdownMenu>
+      <SidebarMenu>
+        <SidebarMenuItem>
+          <DropdownMenuTrigger
+            render={
+              <SidebarMenuButton
+                size="lg"
+                className="flex items-center gap-2.5 rounded-lg px-2 data-[state=open]:bg-sidebar-accent"
+                aria-label="User menu"
+              />
+            }
+          >
+            <div className="flex size-8 shrink-0 items-center justify-center rounded-full bg-brand-soft">
+              <span className="text-xs font-semibold text-brand">{initial}</span>
+            </div>
+            <div className="grid flex-1 text-left text-sm leading-tight group-data-[collapsible=icon]:hidden">
+              <span className="truncate font-medium">{user?.fullName || "Account"}</span>
+              <span className="truncate text-xs text-muted-foreground">
+                {user?.primaryEmailAddress?.emailAddress || ""}
+              </span>
+            </div>
+            <ChevronsUpDownIcon className="ml-auto size-4 shrink-0 group-data-[collapsible=icon]:hidden" />
+          </DropdownMenuTrigger>
+        </SidebarMenuItem>
+      </SidebarMenu>
+      <DropdownMenuContent side={isMobile ? "bottom" : "right"} align="end" className="w-56">
+        <DropdownMenuLabel>
+          <div className="flex flex-col gap-0.5">
+            <span className="text-sm font-medium">{user?.fullName || "Account"}</span>
+            <span className="text-xs text-muted-foreground">
+              {user?.primaryEmailAddress?.emailAddress}
+            </span>
+          </div>
+        </DropdownMenuLabel>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem render={<Link href="/dashboard" />}>
+          <UserIcon className="size-4" />
+          Profile
+        </DropdownMenuItem>
+        <DropdownMenuItem render={<Link href="/dashboard/settings/organizations" />}>
+          <SettingsIcon className="size-4" />
+          Settings
+        </DropdownMenuItem>
+        <DropdownMenuSeparator />
+        <div className="flex items-center justify-between px-2 py-1.5">
+          <span className="text-sm">Theme</span>
+          <ThemeToggle size="h-3.5 w-3.5" className="size-8" />
+        </div>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem
+          className="text-destructive focus:text-destructive"
+          onClick={() => signOut()}
+        >
+          <LogOutIcon className="size-4" />
+          Sign out
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
